Area and Volume Calculator
Gives you a prompted menu to pick an object and enter the appropriate measurements.  It then calculates and displays the total area or volume.  Prompts you to press any key to continue.  I plan on working on a GUI for this to make it a bit more user friendly and expand the shapes you can calculate.